+++ saleor/static/dashboard-next/components/FilterBar/index.ts
@@ -1,0 +1,2 @@
+export { default } from './FilterBar';
+export * from './FilterBar';
