+++ saleor/static/dashboard-next/products/components/ProductListCard/ProductListCard.tsx
@@ -10,14 +10,14 @@ import ProductList from "@saleor/components/ProductList";
 import { Filter } from "@saleor/components/TableFilter";
 import { CategoryDetails_category_products_edges_node } from "../../../categories/types/CategoryDetails";
 import i18n from "../../../i18n";
-import { FilterViewProps, ListActions, PageListProps } from "../../../types";
+import { FilterPageProps, ListActions, PageListProps } from "../../../types";
 import { ProductListUrlFilters } from "../../urls";
 import ProductListFilter from "../ProductListFilter";
 
 export interface ProductListCardProps
   extends PageListProps,
     ListActions,
-    FilterViewProps<FilterContentSubmitData, ProductListUrlFilters> {
+    FilterPageProps<ProductListUrlFilters> {
   currencySymbol: string;
   products: CategoryDetails_category_products_edges_node[];
 }
