+++ saleor/static/dashboard-next/products/components/ProductListFilter/ProductListFilter.tsx
@@ -16,10 +16,7 @@ import { FilterProps } from "../../../types";
 import { StockAvailability } from "../../../types/globalTypes";
 import { ProductListUrlFilters } from "../../urls";
 
-interface ProductListFilterProps
-  extends FilterProps<FilterContentSubmitData, ProductListUrlFilters> {
-  currencySymbol: string;
-}
+type ProductListFilterProps = FilterProps<ProductListUrlFilters>;
 
 export enum ProductFilterKeys {
   published,
@@ -100,71 +97,8 @@ const filterMenu: IFilter = [
   }
 ];
 
-const ProductListFilter: React.FC<ProductListFilterProps> = ({
-  allTabLabel,
-  currencySymbol,
-  filterLabel,
-  filtersList,
-  filterTabs,
-  currentTab,
-  initialSearch,
-  searchPlaceholder,
-  onAll,
-  onSearchChange,
-  onFilterAdd,
-  onFilterSave,
-  onTabChange,
-  onFilterDelete
-}) => {
-  const [search, setSearch] = React.useState(initialSearch);
-  React.useEffect(() => setSearch(initialSearch), [currentTab, initialSearch]);
-
-  const isCustom = currentTab === filterTabs.length + 1;
-
-  return (
-    <>
-      <FilterTabs currentTab={currentTab}>
-        <FilterTab label={allTabLabel} onClick={onAll} />
-        {filterTabs.map((tab, tabIndex) => (
-          <FilterTab
-            onClick={() => onTabChange(tabIndex + 1)}
-            label={tab.name}
-          />
-        ))}
-        {isCustom && (
-          <FilterTab
-            onClick={() => undefined}
-            label={i18n.t("Custom Filter")}
-          />
-        )}
-      </FilterTabs>
-      <Debounce debounceFn={onSearchChange}>
-        {debounceSearchChange => {
-          const handleSearchChange = (event: React.ChangeEvent<any>) => {
-            const value = event.target.value;
-            setSearch(value);
-            debounceSearchChange(value);
-          };
-
-          return (
-            <FilterChips
-              currencySymbol={currencySymbol}
-              menu={filterMenu}
-              filtersList={filtersList}
-              filterLabel={filterLabel}
-              placeholder={searchPlaceholder}
-              search={search}
-              onSearchChange={handleSearchChange}
-              onFilterAdd={onFilterAdd}
-              onFilterSave={onFilterSave}
-              isCustomSearch={isCustom}
-              onFilterDelete={onFilterDelete}
-            />
-          );
-        }}
-      </Debounce>
-    </>
-  );
-};
+const ProductListFilter: React.FC<ProductListFilterProps> = props => (
+  <FilterBar {...props} filterMenu={filterMenu} />
+);
 ProductListFilter.displayName = "ProductListFilter";
 export default ProductListFilter;
